runs/copilot@1-iteration_1/engine.py
import sys

class ChessEngine:
    def __init__(self):
        self.board = self.create_initial_board()
        self.current_turn = 'w'
        self.move_history = []
        self.responses = []

    def create_initial_board(self):
        return [
            ['r', 'n', 'b', 'q', 'k', 'b', 'n', 'r'],
            ['p', 'p', 'p', 'p', 'p', 'p', 'p', 'p'],
            ['.', '.', '.', '.', '.', '.', '.', '.'],
            ['.', '.', '.', '.', '.', '.', '.', '.'],
            ['.', '.', '.', '.', '.', '.', '.', '.'],
            ['.', '.', '.', '.', '.', '.', '.', '.'],
            ['P', 'P', 'P', 'P', 'P', 'P', 'P', 'P'],
            ['R', 'N', 'B', 'Q', 'K', 'B', 'N', 'R']
        ]

    def is_valid_move(self, move):
        # Implement move validation logic here
        return True

    def make_move(self, move):
        if self.is_valid_move(move):
            self.move_history.append(move)
            # Update the board with the move
            self.current_turn = 'b' if self.current_turn == 'w' else 'w'
            return True
        return False

    def generate_legal_moves(self):
        # Implement move generation logic here
        return []

    def handle_uci_command(self, command):
        if command == 'uci':
            self.send_response('id name SimpleChessEngine')
            self.send_response('id author CodeGen')
            self.send_response('uciok')
        elif command == 'isready':
            self.send_response('readyok')
        elif command.startswith('position'):
            self.set_position(command)
        elif command.startswith('go'):
            self.go(command)
        elif command == 'ucinewgame':
            self.new_game()
        elif command == 'quit':
            sys.exit()

    def send_response(self, response):
        print(response)
        self.responses.append(response)

    def set_position(self, command):
        # Implement position setting logic here
        pass

    def go(self, command):
        # Implement move calculation logic here
        best_move = 'e2e4'  # Placeholder for the best move
        self.send_response(f'bestmove {best_move}')

    def new_game(self):
        self.board = self.create_initial_board()
        self.current_turn = 'w'
        self.move_history = []

if __name__ == "__main__":
    engine = ChessEngine()
    while True:
        command = input()
        engine.handle_uci_command(command)
